+++ dymension/libs/kaspa/lib/core/src/deposit.rs
@@ -1,7 +1,5 @@
 use super::escrow::*;
 
-use eyre::Error as EyreError;
-
 use bytes::Bytes;
 
 use std::sync::Arc;
@@ -15,19 +13,37 @@ use kaspa_wallet_core::prelude::*;
 
 use workflow_core::abortable::Abortable;
 
-pub struct DepositFXG;
+use hyperlane_core::HyperlaneMessage;
+use hyperlane_core::H256;
+use serde::{Serialize, Deserialize};
+
+#[derive(Debug, PartialEq, Serialize, Deserialize)]
+pub struct DepositFXG {
+    pub msg_id: H256,
+    pub tx_id: String,
+    pub utxo_index: usize,
+    pub block_id: String,
+    pub payload: HyperlaneMessage,
+}
+
 
 impl TryFrom<Bytes> for DepositFXG {
-    type Error = EyreError;
+    type Error = eyre::Report;
 
     fn try_from(bytes: Bytes) -> Result<Self, Self::Error> {
-        unimplemented!()
+        // Deserialize the bytes into DepositFXG using bincode
+        bincode::deserialize(&bytes)
+            .map_err(|e| eyre::Report::new(e).wrap_err("Failed to deserialize DepositFXG from bytes"))
+
     }
 }
 
 impl From<&DepositFXG> for Bytes {
     fn from(deposit: &DepositFXG) -> Self {
-        unimplemented!()
+        // Serialize the DepositFXG into bytes using bincode
+        let encoded: Vec<u8> = bincode::serialize(deposit)
+            .expect("Failed to serialize DepositFXG into bytes"); 
+        Bytes::from(encoded)
     }
 }
