+++ dymension/libs/kaspa/lib/relayer/src/lib.rs
@@ -14,46 +14,14 @@ use std::error::Error;
 use std::str::FromStr;
 use hyperlane_core::HyperlaneMessage;
 use hyperlane_core::RawHyperlaneMessage;
-use bytes::Bytes;
 use api_rs::apis::{configuration, kaspa_transactions_api::{get_transaction_transactions_transaction_id_get,GetTransactionTransactionsTransactionIdGetParams}};
 use hyperlane_core::U256;
-use hyperlane_core::H256;
 use hyperlane_warp_route::TokenMessage;
 use hyperlane_core::Decode;
 use kaspa_consensus_core::tx::TransactionOutpoint;
 use kaspa_hashes::Hash;
-use serde::{Serialize, Deserialize};
 use anyhow::Result;
-
-
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
-pub struct DepositFXG {
-    pub msg_id: H256,
-    pub tx_id: String,
-    pub utxo_index: usize,
-    pub block_id: String,
-    pub payload: HyperlaneMessage,
-}
-
-impl TryFrom<Bytes> for DepositFXG {
-    type Error = eyre::Report;
-
-    fn try_from(bytes: Bytes) -> Result<Self, Self::Error> {
-        // Deserialize the bytes into DepositFXG using bincode
-        bincode::deserialize(&bytes)
-            .map_err(|e| eyre::Report::new(e).wrap_err("Failed to deserialize DepositFXG from bytes"))
-
-    }
-}
-
-impl From<&DepositFXG> for Bytes {
-    fn from(deposit: &DepositFXG) -> Self {
-        // Serialize the DepositFXG into bytes using bincode
-        let encoded: Vec<u8> = bincode::serialize(deposit)
-            .expect("Failed to serialize DepositFXG into bytes"); 
-        Bytes::from(encoded)
-    }
-}
+use core::deposit::DepositFXG;
 
 
 pub fn parse_hyperlane_message(m: &RawHyperlaneMessage) -> Result<HyperlaneMessage,anyhow::Error> {
@@ -79,64 +47,67 @@ fn get_tn10_config() -> configuration::Configuration {
     }
 }
 
+pub async fn handle_new_deposit(tx: String) -> Result<DepositFXG, Box<dyn Error>> {
+    // rpc config
+    let config = get_tn10_config();
+
+    let get_params = GetTransactionTransactionsTransactionIdGetParams  {
+        transaction_id: tx.clone(),
+        block_hash: None,
+        inputs: None,
+        outputs: None,
+        resolve_previous_outpoints: None,
+    };
+    // get transaction info using Kaspa API
+    let res = get_transaction_transactions_transaction_id_get(&config, get_params).await?;
+    let payload = res.payload.ok_or("Tx payload not found")?;
+    let block_id = res.accepting_block_hash.ok_or("Block id not found")?;
+
+    // decode payload into Hyperlane message
+    let rawmessage: RawHyperlaneMessage = hex::decode(payload)?;
+    let mut message = parse_hyperlane_message(&rawmessage)?;
+
+    // decode token message inside  Hyperlane message
+    let mut reader = Cursor::new(message.body.as_slice());
+    let token_message =  TokenMessage::read_from(&mut reader)?;
+
+    // find the index of the utxo that satisfies the transfer amount in hl message
+    let utxo_index = res.outputs.ok_or("no utxo found in tx")?.iter().position(|utxo: &api_rs::models::TxOutput| U256::from(utxo.amount) >= token_message.amount()).ok_or("no utx found")?;
+
+    // builds the TransactionOutpoint to inject to hl message 
+    let tx_id = res.transaction_id.ok_or("tx id not found")?;
+    let tx_hash = Hash::from_str(&tx_id)?;
+    let output = TransactionOutpoint{
+        transaction_id: tx_hash,
+        index: utxo_index as u32,
+    };
+    let output_bytes = bincode::serialize(&output)?;
+
+    // replace kaspa value and reencode message
+    let mut metadata: HlMetadata = HlMetadata::decode(token_message.metadata())?;
+    metadata.kaspa = output_bytes;
+    let body = metadata.encode_to_vec();
+    message.body = body;
+
+    // build response for validator
+    let tx  = DepositFXG{
+        msg_id: message.id(),
+        tx_id: tx,
+        utxo_index: utxo_index,
+        block_id: block_id,
+        payload: message,
+    };
+    Ok(tx)
+}
 
 pub async fn handle_new_deposits(
     transaction_ids: Vec<String>,
 ) -> Result<Vec<DepositFXG>, Box<dyn Error>> {
 
-    // rpc config
-    let config = get_tn10_config();
-
     let mut txs = Vec::new(); 
 
     for transaction in transaction_ids {
-
-        let get_params = GetTransactionTransactionsTransactionIdGetParams  {
-            transaction_id: transaction.clone(),
-            block_hash: None,
-            inputs: None,
-            outputs: None,
-            resolve_previous_outpoints: None,
-        };
-        // get transaction info using Kaspa API
-        let res = get_transaction_transactions_transaction_id_get(&config, get_params).await?;
-        let payload = res.payload.ok_or("Tx payload not found")?;
-        let block_id = res.accepting_block_hash.ok_or("Block id not found")?;
-
-        // decode payload into Hyperlane message
-        let rawmessage: RawHyperlaneMessage = hex::decode(payload)?;
-        let mut message = parse_hyperlane_message(&rawmessage)?;
-
-        // decode token message inside  Hyperlane message
-        let mut reader = Cursor::new(message.body.as_slice());
-        let token_message =  TokenMessage::read_from(&mut reader)?;
-
-        // find the index of the utxo that satisfies the transfer amount in hl message
-        let utxo_index = res.outputs.ok_or("no utxo found in tx")?.iter().position(|utxo: &api_rs::models::TxOutput| U256::from(utxo.amount) >= token_message.amount()).ok_or("no utx found")?;
-
-        // builds the TransactionOutpoint to inject to hl message 
-        let tx_id = res.transaction_id.ok_or("tx id not found")?;
-        let tx_hash = Hash::from_str(&tx_id)?;
-        let output = TransactionOutpoint{
-            transaction_id: tx_hash,
-            index: utxo_index as u32,
-        };
-        let output_bytes = bincode::serialize(&output)?;
-
-        // replace kaspa value and reencode message
-        let mut metadata: HlMetadata = HlMetadata::decode(token_message.metadata())?;
-        metadata.kaspa = output_bytes;
-        let body = metadata.encode_to_vec();
-        message.body = body;
-
-        // build response for validator
-        let tx  = DepositFXG{
-            msg_id: message.id(),
-            tx_id: transaction,
-            utxo_index: utxo_index,
-            block_id: block_id,
-            payload: message,
-        };
+        let tx = handle_new_deposit(transaction).await?;
         txs.push(tx);
 
     }
